+++ src/index.ts
@@ -1,3 +1,16 @@
 import FragmentPlugin from './SwupFragmentPlugin.js';
 export default FragmentPlugin;
-export type { Options, Rule, FragmentElement } from './inc/types.js';
+import type { Route, FragmentVisit } from './inc/defs.js';
+export type { Options, Rule, FragmentElement, FragmentVisit } from './inc/defs.js';
+
+declare module 'swup' {
+	export interface Swup {
+		getFragmentVisit?: (route: Route) => FragmentVisit | undefined;
+	}
+	export interface Visit {
+		fragmentVisit?: FragmentVisit;
+	}
+	export interface CacheData {
+		fragmentHtml?: string;
+	}
+}
